Updated sheet window inheritance everywhere, and re-ordered some of them into a single inspector file
