+++ src/main/resources/base/config.ini
@@ -10,9 +10,9 @@ medianPearsonCorr = False
 lagsPearsonCorr = 30
 donorLeft = False
 hmmBICStrictness = 2.0
-lastOpenedDir = /Users/Joh/Desktop/fits traces
-contrastBoxHiGrnVal = 100.0
-contrastBoxHiRedVal = 100.0
+lastOpenedDir = /Users/Joh/Desktop
+contrastBoxHiGrnVal = 50.0
+contrastBoxHiRedVal = 50.0
 alphaFactor = 0.0
 deltaFactor = 0.0
 betaFactor = 0
